requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.2
Django==3.2.6
djangorestframework==3.12.4
filelock==3.0.12
geographiclib==1.52
geopy==2.2.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
pipenv==2021.5.29
platformdirs==2.2.0
psycopg2-binary==2.9.1
pyTelegramBotAPI==3.8.3
pytz==2021.1
PyYAML==5.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.7.2
virtualenv-clone==0.5.6
